requirements.txt
attrs==23.1.0
certifi==2023.7.22
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
numpy==1.25.2
outcome==1.2.0
pandas==2.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
